% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_min_dist.R
\name{distancia}
\alias{distancia}
\title{Distancia entre coordenadas}
\usage{
distancia(x, y)
}
\arguments{
\item{x}{Local de partida}

\item{y}{Local de chegada}
}
\value{
Distancia entre x e y
}
\description{
Distancia entre coordenadas
}
