% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_geopart.R
\name{geopart}
\alias{geopart}
\title{geopart}
\usage{
geopart(
  shape_file,
  dataset,
  coord_names = c("lon", "lat"),
  proj_specs = "+proj=longlat +zone=23 +south +ellps=aust_SA +units=m +no_defs"
)
}
\arguments{
\item{shape_file}{A SpatialPolygonsDataFrame object created using sp package}

\item{dataset}{A data.frame or matrix object containing lon/lat coordinates}

\item{coord_names}{List of characters contaning variable names for lon/lat collumns. Defaults to c("lon","lat")}
}
\description{
Partitions a dataset according to the location area each observation belongs to.
}
