% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_addr.R
\name{get_addr}
\alias{get_addr}
\title{get_addr}
\usage{
get_addr(address = NULL)
}
\arguments{
\item{address}{Character value containing street adress}
}
\description{
Returns the corresponding geocoordinates (latitude, longitude) of a given an address .
Uses Open Street Maps JSON API (requires jsonlite). By D.Kisler
}
