% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipvs.R
\name{ler_ipvs}
\alias{ler_ipvs}
\title{Ler dados do Indice Paulista de Vulnerabilidade Social (IPVS)}
\usage{
ler_ipvs()
}
\value{
Uma tibble
}
\description{
Ler dados do Indice Paulista de Vulnerabilidade Social (IPVS)
}
