% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist.R
\name{min_dist}
\alias{min_dist}
\title{min_dist}
\usage{
min_dist(coords, targets)
}
\arguments{
\item{coords}{A nx2 matrix or data frame containing geocoordinates for n subjects.}

\item{targets}{A mx2 matrix containing a set of geocoordinates for m target locations.}
}
\description{
Returns the minimum distance to a set of targets given a dataframe with coordinates.
Uses sp spDistsN1 with Great Circle distance as default
}
