% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_nearplace.R
\name{nearplace}
\alias{nearplace}
\title{nearplace}
\usage{
nearplace(
  coords,
  targets,
  coord_names = c("lon", "lat"),
  great_circ = T,
  target_label = "names"
)
}
\arguments{
\item{coords}{A nx2 matrix or data frame containing geocoordinates for n subjects.}

\item{points}{A mx2 matrix containing a set of geocoordinates for m target locations.}
}
\description{
Returns the nearest pleace given a set of targets and a dataframe with coordinates.
Uses sp spDistsN1 with Great Circle distance as default.
}
