% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_min_dist.R
\name{remove_zero}
\alias{remove_zero}
\title{Remove distancias a si mesmo das matrizes}
\usage{
remove_zero(X)
}
\arguments{
\item{X}{Matriz}
}
\value{
Matriz sem zeros
}
\description{
Remove distancias a si mesmo das matrizes
}
