% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopart.R
\name{geopart}
\alias{geopart}
\title{geopart}
\usage{
geopart(dataset, shape_file)
}
\arguments{
\item{dataset}{A data.frame or matrix object containing lon/lat coordinates}

\item{shape_file}{A SpatialPolygonsDataFrame object created using sp package}

\item{coord_names}{List of characters contaning variable names for lon/lat collumns. Defaults to c("lon","lat")}
}
\description{
Partitions a dataset according to the location area each observation belongs to.
}
